+++ src/components/SortableTable.js
@@ -1,40 +1,21 @@
-import { useState } from 'react';
 import Table from './Table';
 import { GoArrowSmallDown, GoArrowSmallUp } from 'react-icons/go'
+import useSort from '../hooks/use-sort';
 
 function SortableTable(props) {
-    const [sortOrder, setSortOrder] = useState(null);
-    const [sortBy, setSortBy] = useState(null)
     const { config, data } = props;
-
-    const handleClick = (label) => {
-        if (sortBy && label !== sortBy) {
-            setSortOrder('asc');
-            setSortBy(label);
-            return;
-        }
-
-        if (sortOrder === null) {
-            setSortOrder('asc');
-            setSortBy(label);
-        } else if (sortOrder === 'asc') {
-            setSortOrder('desc')
-            setSortBy(label);
-        } else if (sortOrder === 'desc') {
-            setSortOrder(null)
-            setSortBy(null)
-        }
-    }
+    
+    const { sortOrder, sortBy, sortedData, setSortColumn } = useSort(data, config)
 
     const updatedConfig = config.map((column) => {
         if (!column.sortValue) {
-        return column;
+            return column;
         }
 
         return {
             ...column,
             header: () => 
-                <th className='cursor-pointer hover:bg-gray-100' onClick={() => handleClick(column.label)}>
+                <th className='cursor-pointer hover:bg-gray-100' onClick={() => setSortColumn(column.label)}>
                     <div className='flex items-center'>
                         {getIcons(column.label, sortBy, sortOrder)}
                         {column.label}
@@ -42,29 +23,6 @@ function SortableTable(props) {
                 </th>
         };
     });
-
-    // Only sort data if sortOrder && sortBy are not null
-    // Make a copy of the data prop
-    // Find the correct sort value function and use it to sort
-
-    let sortedData = data
-
-    if (sortOrder && sortBy) {
-        const { sortValue } = config.find(column => column.label === sortBy)
-        
-        sortedData = [...data].sort((a, b) => {
-            const valueA = sortValue(a);
-            const valueB = sortValue(b);
-
-            const reverseOrder = sortOrder === 'asc' ? 1 : -1;
-
-            if (typeof valueA === 'string') {
-                return valueA.localeCompare(valueB) * reverseOrder;
-            } else {
-                return (valueA - valueB) * reverseOrder;
-            }
-        });
-    }
 
     return <Table {...props} data={sortedData} config={updatedConfig} />
 }
